some cleanup
- removed content menu and content menu item from interfaces
- added content menu item to content menu
- removed private var and public getter
- removed some extra console logs
